Create MontyHallProblem.pl
Initial Perl script to explore the results of changing or staying with an initial selection in the Monty Hall Problem. This script allows for several user options that demonstrate how the results of the problem change under various conditions.

Run with --help to get a list of available user options.